Move all helper functions from sai_base_test to sai_utils
